Expanded functions added to LMS see description
For Module 6 - Software Implementation Phase 1 - Logic and Functionality, expanded functionality has been implemented allowing users to check in and out books with the barcode number. The name "id" has been completely replaced with a more fitting name "barcode" and when the user lists all books, it now shows whether the book has been checked in or out. Upon trying to check out a book that has already been checked out, the user will be met with the appropriate response. Same thing if the user enters a non numerical option or just a wrong number. This should prevent infinite loops and program crashes.