Module 7 - Software Testing Assignment Commit
For Module 7 - Software Testing Assignment, we were instructed to use junit testing to run a variety of unit tests to verify that books can be added, removed and checked out or in. I attempted to set up JUnit in my project in IntelliJ, but met some difficulty. Working with the tutorial provided I was able to get my test set up but unfortunately not able to identify the classes in my project to successfully run the tests. I will need to continue working with this and find a solution.