Module 8 Software Implementation Phase 2  Commit
For this commit, the GUI has been implemented for the LMS. By running "LibraryManagementSystemGUI", the user will be presented the interface that allows them to navigate through the program. When selecting "Add Books", enter in the box: Books.txt, and the books from that file will be added and displayed on the text box below.

I checked for any possible errors, if the user enters an incorrect integer or non numerical character, they will be presented with the appropriate response.