Submission for Module 9
For the Module 9 submission, I updated all classes to now pull from a database instead of a .txt file. I attempted to get the LMS connected to the library.db file that is located in the root directory, but was unfortunately was unsuccessful. I will need to continue to check my file directory for the database and see what may be causing the LMS to not load from the database file included.